Fix float currency

+++ economy-bot/classes/economy_bot.py
@@ -269,7 +269,7 @@ class EconomyBot(discord.ext.commands.Bot):
                         f"User {registered_user.user_id} wrote {messages} message in the last {configuration.check_timer} minutes")
                     if messages >= int(configuration.check_minimum_messages):
                         rate = messages / int(configuration.check_maximum_messages)
-                        reward = rate * int(configuration.check_maximum_currency)
+                        reward = float(rate * int(configuration.check_maximum_currency))
                         await database.automatic_reward(user_id=registered_user.user_id, guild_id=guild.id,
                                                         amount=reward)
